upload_static: add script
This script will upload static files to S3 bucket. It uses some private
functions from flask_s3 because flask_s3.create_all will try to create
the bucket, and we don't have permission to create the bucket, just to
use it. We may send a patch to them later.

Related to #6. Once I add a hook for uploading files, we can close this
issue.